Complete Calculation OK. Change Completed Fields To Float.

+++ EmpiriaBMS.Core/Dtos/DisciplineDto.cs
@@ -17,7 +17,7 @@ public class DisciplineDto : EntityDto
 
     public long MenHours { get; set; }
 
-    public int Completed { get; set; }
+    public float Completed { get; set; }
 
     public int? EngineerId { get; set; }
     public User? Engineer { get; set; }
